backend/requirements.txt
accelerate>=1,<2
diffusers>=0.31,<1
fastapi==0.115.6
numpy>=1.26,<2
Pillow==11.1.0
safetensors>=0.4,<1
transformers>=4.45,<5
uvicorn==0.32.1
